ai-service/app/services/photo_processor.py
from PIL import Image
from io import BytesIO
import logging

logger = logging.getLogger(__name__)

class PhotoProcessor:
    def __init__(self):
        pass
    
    def analyze_photo(self, image_data: bytes) -> dict:
        """
        Analyze photo and extract metadata.
        This is a placeholder for actual AI processing.
        """
        try:
            image = Image.open(BytesIO(image_data))
            
            metadata = {
                "width": image.width,
                "height": image.height,
                "format": image.format,
                "mode": image.mode,
                "tags": self._generate_tags(image)
            }
            
            logger.info(f"Photo analyzed: {metadata}")
            return metadata
        except Exception as e:
            logger.error(f"Error analyzing photo: {e}")
            raise
    
    def _generate_tags(self, image: Image.Image) -> list[str]:
        """
        Generate tags based on image analysis.
        This is a simple placeholder - in production, you'd use ML models.
        """
        tags = []
        
        # Basic analysis
        if image.width > image.height:
            tags.append("landscape")
        elif image.height > image.width:
            tags.append("portrait")
        else:
            tags.append("square")
        
        # Color mode
        if image.mode == "RGB":
            tags.append("color")
        elif image.mode in ["L", "1"]:
            tags.append("grayscale")
        
        # Size category
        total_pixels = image.width * image.height
        if total_pixels > 2000000:  # > 2MP
            tags.append("high-resolution")
        else:
            tags.append("standard-resolution")
        
        return tags

photo_processor = PhotoProcessor()
