ai-service/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
redis==5.0.1
Pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
